usg-scripts/requirements.txt
azure.storage.queue
